### 2025-02-09 #7762 @editorialbot
[PRE REVIEW]: embedplyr: Tools for Working With Text Embeddings
**Submitting author:** @rimonim (<a href="http://orcid.org/0009-0001-9347-0145">Louis Teitelbaum</a>)
**Repository:** https://github.com/rimonim/embedplyr
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
