### 2025-02-10 a19fb69
optional standard errors for total_dist
