+++ paper/paper.md
@@ -17,16 +17,14 @@ affiliations:
 date: 10 January 2025
 year: 2025
 bibliography: paper.bib
-output:
-  html_document:
-    df_print: paged
+output: rticles::joss_article
 csl: apa.csl
 journal: JOSS
 ---
 
 
 
-# Aims of the Package
+# Statement of Need
 
 Dense vector embeddings are the fundamental building block of modern natural language processing [@lauriola2022]. The ability to represent the meaning of texts as a continuous, high dimensional space underlies the recent successes of large language models (LLMs). Embeddings have also revolutionized research methods and inspired new theoretical frameworks in linguistics, psychology, sociology, and neuroscience [see e.g. @duran2019; @feuerriegel2025; @grand2022; @hamilton2018; @kjell2023; @kozlowski2019; @schrimpf2021]. As text embeddings become ubiquitous in the social and behavioral sciences, the need for flexible, easy-to-learn tools increases. Answering this need, **embedplyr** (pronounced "embe-DEE-plier") enables common operations with word and text embeddings within a **tidyverse** [@wickham2019] and/or **quanteda** [@benoit2018] workflow, as demonstrated in @teitelbaum2024. **embedplyr** is designed to facilitate the use of word and text embeddings in common data manipulation and text analysis workflows, without introducing new syntax or unfamiliar data structures to R users.
 
@@ -38,7 +36,7 @@ Existing tools for working with embeddings in R are generally specific to partic
 
 **embedplyr** does not include tools for training new embedding models, but it can load embeddings from a file or download them from online. This is especially useful for pretrained word embedding models like GloVe [@pennington2014], word2vec [@mikolov2013a; @mikolov2013b], HistWords [@hamilton2018], and fastText [@bojanowski2017a]. Hundreds of these models can be conveniently downloaded from online sources with `load_embeddings()`, forgoing the need to search for model files online and juggle incompatible file types. 
 
-One particularly useful feature of `load_embeddings()` is the optional `words` parameter, which allows the user to specify a subset of words to load from the model. This allows users to work with large models, which are often too large to load into an interactive environment in their entirety. For example, a user may specify the set of unique tokens in their corpus of interest, and load only these from the model.
+One particularly useful feature of `load_embeddings()` is the optional `words` parameter, which allows the user to specify a subset of words to load from the model. This allows users to work with large models, which are often too large to load into an interactive environment in their entirety. 
 
 
 ```r
@@ -96,7 +94,7 @@ Whereas indexing a regular matrix by rownames gets slower as the number of rows 
 
 Given a tidy dataframe of texts, `embed_docs()` will generate embeddings by averaging the embeddings of tokens in each text [see @ethayarajh2019; @teitelbaum2024, chap. 18]. By default, `embed_docs()` uses a simple unweighted mean, but many other averaging methods are available.
 
-The following example embeds three texts, which for the sake of the example I will consider to have been written by a participant diagnosed with depression, one diagnosed with anxiety, and one control. 
+The following example embeds three texts, which for the sake of the example will can considered to have been written by one participant diagnosed with depression, one diagnosed with anxiety, and one control. 
 
 
 ```r
@@ -118,11 +116,9 @@ psych_embeddings_df <- psych_df |>
 
 ## Embed Dictionaries
 
[…]
