+++ paper/paper.bib
@@ -14,6 +14,7 @@
 	volume = {24},
 	copyright = {2019 American Psychological Association},
 	issn = {1082-989X},
+	doi = {10.31234/osf.io/a5yh9},
 	abstract = {Linguistic alignment (LA) is the tendency during a conversation to reuse each other's linguistic expressions, including lexical, conceptual, or syntactic structures. LA is often argued to be a crucial driver in reciprocal understanding and interpersonal rapport, though its precise dynamics and effects are still controversial. One barrier to more systematic investigation of these effects lies in the diversity in the methods employed to analyze LA, which makes it difficult to integrate and compare results of individual studies. To overcome this issue, we have developed ALIGN (Analyzing Linguistic Interactions with Generalizable techNiques), an open-source Python package to measure LA in conversation (https://pypi.python.org/pypi/align) along with in-depth open-source tutorials hosted on ALIGN's GitHub repository (https://github.com/nickduran/align-linguistic-alignment). Here, we first describe the challenges in the study of LA and outline how ALIGN can address them. We then demonstrate how our analytical protocol can be applied to theory-driven questions using a complex corpus of dialogue (the Devil's Advocate corpus; Duran \& Fusaroli, 2017). We close by identifying further challenges and point to future developments of the field.},
 	number = {4},
 	journal = {Psychological methods},
@@ -156,7 +157,7 @@ Publisher: American Psychological Association},
 	title = {A generalized solution of the orthogonal procrustes problem},
 	volume = {31},
 	issn = {1860-0980},
-	url = {https://doi.org/10.1007/BF02289451},
+	url = {10.1007/BF02289451},
 	doi = {10.1007/BF02289451},
 	abstract = {A solutionT of the least-squares problemAT=B +E, givenA andB so that trace (E′E)= minimum andT′T=I is presented. It is compared with a less general solution of the same problem which was given by Green [5]. The present solution, in contrast to Green's, is applicable to matricesA andB which are of less than full column rank. Some technical suggestions for the numerical computation ofT and an illustrative example are given.},
 	language = {en},
@@ -175,6 +176,7 @@ Publisher: American Psychological Association},
 	url = {https://psychbruce.github.io/PsychWordVec/},
 	author = {Bao, Han-Wu-Shuang},
 	year = {2023},
+	doi = {10.32614/cran.package.psychwordvec},
 	annote = {R package version 2023.9},
 }
 
@@ -368,6 +370,7 @@ The book is published at ds4psych.com},
 	copyright = {Apache License (≥ 2.0)},
 	shorttitle = {word2vec},
 	url = {https://cran.r-project.org/web/packages/word2vec/index.html},
+	doi = {10.32614/cran.package.word2vec},
 	abstract = {Learn vector representations of words by continuous bag of words and skip-gram implementations of the 'word2vec' algorithm. The techniques are detailed in the paper "Distributed Representations of Words and Phrases and their Compositionality" by Mikolov et al. (2013), available at {\textless}doi:10.48550/arXiv.1310.4546{\textgreater}.},
 	urldate = {2025-01-05},
 	author = {Wijffels, Jan and Watanabe, Kohei and Fomichev, Max},
@@ -381,6 +384,7 @@ The book is published at ds4psych.com},
 	copyright = {GPL-2 {\textbar} GPL-3 {\textbar} file LICENSE [expanded from: GPL (≥ 2) {\textbar} file LICENSE]},
 	shorttitle = {text2vec},
 	url = {https://cran.r-project.org/web/packages/text2vec/index.html},
+	doi = {10.32614/cran.package.text2vec},
[…]
