### 2025-02-11 #7762 @rimonim
Paper has been shortened and references fixed.
Suggestions for potential reviewers: Christopher Barrie, Kenneth Benoit, Justin Chun-ting Ho, Kenneth Enevoldsen
