### 2025-02-11 046b261
optional standard errors for total_dist
