### 2025-03-18 #7762 @britta-wstnr
Hi @rimonim!
Could you have a look at the invalid DOI reported by our trusty editorial bot [here](https://github.com/openjournals/joss-reviews/issues/7762#issuecomment-2734557576)?
I am reaching out to potential reviewers via email here, as am I not sure how active they are on Github. I will update you periodically on how this is going! If you have any questions, you can of course ping me here any time.
