+++ paper/paper.bib
@@ -397,7 +397,7 @@ The book is published at ds4psych.com},
 	title = {Latent dirichlet allocation},
 	volume = {3},
 	issn = {1532-4435},
-	doi = {https://doi.org/10.7551/mitpress/1120.003.0082},
+	doi = {10.7551/mitpress/1120.003.0082},
 	abstract = {We describe latent Dirichlet allocation (LDA), a generative probabilistic model for collections of discrete data such as text corpora. LDA is a three-level hierarchical Bayesian model, in which each item of a collection is modeled as a finite mixture over an underlying set of topics. Each topic is, in turn, modeled as an infinite mixture over an underlying set of topic probabilities. In the context of text modeling, the topic probabilities provide an explicit representation of a document. We present efficient approximate inference techniques based on variational methods and an EM algorithm for empirical Bayes parameter estimation. We report results in document modeling, text classification, and collaborative filtering, comparing to a mixture of unigrams model and the probabilistic LSI model.},
 	number = {null},
 	journal = {J. Mach. Learn. Res.},
