### 2025-06-24 rimonim/embedplyr#1 @cmaimone
Contributing guidelines
Re: https://github.com/openjournals/joss-reviews/issues/8262

I see a code of conduct, but I'm not seeing contributing guidelines (I may be missing them...)
