### 2025-06-24 rimonim/embedplyr#2 @cmaimone
JOSS paper performance claim
There's a mild performance claim in the paper: token embeddings can be retrieved in milliseconds even from models with millions of rows.

Is there a test or analysis to support this somewhere in the repo?

Re: https://github.com/openjournals/joss-reviews/issues/8262
