### 2025-06-24 rimonim/embedplyr#4 @cmaimone
load_embeddings help page
There is a note for some GloVe models on the [load_embeddings](https://rimonim.github.io/embedplyr/reference/load_embeddings.html) documentation that notes that downloading just selected words doesn't work - that the whole model needs to be downloaded.  I think that applies to more models than just the few that have that note.  Example:

```
> esub <- load_embeddings(
+   "glove.6B.50d",
+   words = c("science", "politics", "country")
+ )
Downloading from the Internet...
trying URL 'https://github.com/piskvorky/gensim-data/releases/download/glove-wiki-gigaword-50/glove-wiki-gigaword-50.gz'
Content type 'application/octet-stream' length 69182535 bytes (66.0 MB)
==================================================
downloaded 66.0 MB

Reading model from file...
Warning message:
In load_embeddings("glove.6B.50d", words = c("science", "politics",  :
  Saving only a subset of words is not supported for original format. Saving full model file.
```

Re: https://github.com/openjournals/joss-reviews/issues/8262
