### 2025-06-24 rimonim/embedplyr#6 @cmaimone
Question: overloading of "predict"
I'm curious about the choice to use predict() for the embeddings object for the case of retrieving specific vectors from the embeddings object.  I can see how it sort of fits, but it is closer to retrieving the results of the model than predicting for new values.  I found it confusing, so I'm wondering if there is some information about what using predict() for this allows someone to do that they couldn't if this was a different function or something to explain why its predict() for this task.

This isn't a requirement for the JOSS review

Re: https://github.com/openjournals/joss-reviews/issues/8262
