### 2025-06-24 rimonim/embedplyr#10 @cmaimone
Question: package name
I'm a little bit confused about the package name.  I very much appreciate being able to work with embeddings in a piped workflow, but I question whether this really matches well with dplyr, since I can't use the core dplyr functions on an embeddings object (directly - I know it can be coerced into a data frame).  Because the embeddings object doesn't have names (column names), I can't use select, filter, etc. from dplyr.  I realize this would be a big change, but I'd consider whether the current package name is the right name.  If you think it is the right name, it might help to explain that a bit more, because just being pipe-friendly is a fairly loose link, especially now that there is a native pipe in R.

Re: https://github.com/openjournals/joss-reviews/issues/8262
