### 2025-06-24 #8262 @cmaimone
Seems like a useful package - I plan to use it next time I want to work with word vectors.

There are a few small issues to address to close out the review from my perspective.  They appear above.  I've added some additional suggestions as issues (generally labeled as such) that aren't always linked directly to this review issue -- those are just ideas for improvements/enhancements going forward and do not bear on the JOSS review.
