### 2025-06-24 rimonim/embedplyr#1 @orhunulusahin
I was conflicted about this. I suppose the use of GPL v3 has some implications for contribution, and the code of conduct mentioned above add to it as well, but it's probably best to have a separate statement about contributing.
