### 2025-08-05 4985bef
performance vignette
addresses issue #2
