+++ paper/paper.Rmd
@@ -64,7 +64,7 @@ moral_embeddings
 find_nearest(glove_twitter_25d, "dog", 5L, method = "cosine")
 ```
 
-Whereas indexing a regular matrix by rownames gets slower as the number of rows increases, **embedplyr**'s hash table indexing means that token embeddings can be retrieved in milliseconds even from models with millions of rows.
+Whereas indexing a regular matrix by rownames gets slower as the number of rows increases, **embedplyr**'s hash table indexing means that token embeddings can be retrieved in milliseconds even from models with millions of rows (see the [performance vignette](vignettes/performance.html)).
 
 ## Embed Texts of Interest
 
