### 2025-08-05 8627020
add file download sizes
addresses issue #3
