+++ man/load_embeddings.Rd
@@ -35,24 +35,25 @@ the model is retrieved from online sources and, by default, saved.
 }
 \details{
 The following are supported models for download. Note that some models are very large.
+Listed file sizes are those of the full file to be downloaded.
 If you know in advance which word embeddings you will need (e.g. the set of unique
 tokens in your corpus), consider specifying this with the \code{words} parameter to save
 memory and processing time.
 \subsection{GloVe}{
 
 \itemize{
-\item \code{glove.42B.300d}: Common Crawl (42B tokens, 1.9M vocab, uncased, 300d). Downloaded from https://huggingface.co/stanfordnlp/glove.
+\item \code{glove.42B.300d}: Common Crawl (42B tokens, 1.9M vocab, uncased, 300d, 1.88 GB). Downloaded from https://huggingface.co/stanfordnlp/glove.
 This file is a zip archive and must temporarily be downloaded in its entirety even when \code{words} is specified.
-\item \code{glove.840B.300d}: Common Crawl (840B tokens, 2.2M vocab, cased, 300d). Downloaded from https://huggingface.co/stanfordnlp/glove.
+\item \code{glove.840B.300d}: Common Crawl (840B tokens, 2.2M vocab, cased, 300d, 2.18 GB). Downloaded from https://huggingface.co/stanfordnlp/glove.
 This file is a zip archive and must temporarily be downloaded in its entirety even when \code{words} is specified.
-\item \code{glove.6B.50d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 50d). Downloaded from https://github.com/piskvorky/gensim-data
-\item \code{glove.6B.100d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 100d). Downloaded from https://github.com/piskvorky/gensim-data
-\item \code{glove.6B.200d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 200d). Downloaded from https://github.com/piskvorky/gensim-data
-\item \code{glove.6B.300d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 300d). Downloaded from https://github.com/piskvorky/gensim-data
-\item \code{glove.twitter.27B.25d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 25d). Downloaded from https://github.com/piskvorky/gensim-data
-\item \code{glove.twitter.27B.50d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 50d). Downloaded from https://github.com/piskvorky/gensim-data
-\item \code{glove.twitter.27B.100d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 100d). Downloaded from https://github.com/piskvorky/gensim-data
-\item \code{glove.twitter.27B.200d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 200d). Downloaded from https://github.com/piskvorky/gensim-data
+\item \code{glove.6B.50d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 50d, 862 MB). Downloaded from https://github.com/piskvorky/gensim-data
+\item \code{glove.6B.100d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 100d, 862 MB). Downloaded from https://github.com/piskvorky/gensim-data
[…]
