### 2025-08-05 d1995db
clarify saving behavior with `format = "original"`
addresses issue #4
