### 2025-08-05 rimonim/embedplyr#4 @rimonim
@cmaimone thank you for bringing this up - this was definitely a weak point in the documentation.
The issue here is that `load_embeddings()` defaults to `format = "original"`, but the package does not include utilities for _writing_ to the various formats that word embedding models come in. So it can handle _reading_ select words from models like the one in your example, but when told to save in the original format it has no choice but to download the whole thing. If you set`format` to `"csv"` or `"rds"` your example should work fine.

I've updated the documentation to clarify this behavior, and added language to the warning pointing users to the solution: "Saving only a subset of words is not supported for original format. Saving full model file. To save only a subset of words, set `format` to 'csv' or 'rds'."

I initially considered changing the default `format` for this reason, but decided against it because the `words` functionality is primarily designed to avoid loading large models into the interactive environment, not to save space on the drive. Even if a user wants to load only a few words into the session, they may still want the full model on their computer for future sessions.
