### 2025-08-06 rimonim/embedplyr#6 @rimonim
Thanks for bringing this up - this is definitely a question worth thinking about. The beginnings of embedplyr were as a few scripts that expanded on the [word2vec](https://cran.r-project.org/web/packages/word2vec/index.html) package, which uses `predict()`, and I never bothered to rethink it. In its defence, e.g. [`predict.lm()`](https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/predict.lm) returns fitted values by default - pretrained embeddings are kind of like fitted values if you don't think about it too hard (if you do think about it too hard, `coef()` would probably be the appropriate function).
That said, let's consider some other options:

1. Get rid of `predict()` entirely and just rely on subsetting, as in [PsychWordVec](https://psychbruce.github.io/PsychWordVec/index.html) and [text2vec](https://text2vec.org/index.html). As noted in the documentation, the only difference between `embeddings[i,]` and `predict(embeddings, i)` is that the former throws an error when items of `i` are not valid indices, whereas the latter handles it gracefully. Possibly this functionality could be worked in to subsetting? But removing the explicit function might make it less discoverable for new users.
2. Name the function something else as in [fastTextR](https://cran.r-project.org/web/packages/fastTextR/vignettes/Word_representations.html)'s `ft_word_vectors()`. Since this is such a basic functionality, I would lean toward something short - maybe `emb()`? Any suggestions?
