### 2025-08-06 27246f8
add links to quanteda documentation
partially addresses issue #8
