### 2025-08-06 4319cda
rdname change from textstat_embedding to embed_docs
partially addresses issue #8
