+++ man/embed_docs.Rd
@@ -1,20 +1,12 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textstat_embedding.R
-\name{textstat_embedding}
-\alias{textstat_embedding}
+% Please edit documentation in R/embed_docs.R
+\name{embed_docs}
 \alias{embed_docs}
+\alias{textstat_embedding}
 \alias{embed_docs.default}
 \alias{embed_docs.data.frame}
 \title{Get Text Embeddings by Averaging Word Embeddings}
 \usage{
-textstat_embedding(
-  dfm,
-  model,
-  w = NULL,
-  method = "mean",
-  output_embeddings = FALSE
-)
-
 embed_docs(x, ...)
 
 \method{embed_docs}{default}(
@@ -39,9 +31,20 @@ embed_docs(x, ...)
   tolower = TRUE,
   output_embeddings = FALSE
 )
+
+textstat_embedding(
+  dfm,
+  model,
+  w = NULL,
+  method = "mean",
+  output_embeddings = FALSE
+)
 }
 \arguments{
-\item{dfm}{a quanteda \link[quanteda:dfm]{dfm}}
+\item{x}{a character vector, a data frame, or data frame extension (e.g. a tibble)}
+
+\item{...}{additional parameters to pass to \code{quanteda::tokens()} or to the
+user-specified modeling function}
 
 \item{model}{an \link{embeddings} object. For \code{embed_docs()}, \code{model}
 can alternatively be a function that takes a character vector and outputs a
@@ -54,22 +57,19 @@ embeddings object. See \code{\link[=average_embedding]{average_embedding()}}.}
 \code{method = "median"} does not use matrix operations and may therefore be slow
 for datasets with many documents.}
 
-\item{output_embeddings}{\code{FALSE} (the default) returns a tibble.
-\code{TRUE} returns an embeddings object. See 'Value' for details.}
-
-\item{x}{a character vector, a data frame, or data frame extension (e.g. a tibble)}
-
-\item{...}{additional parameters to pass to \code{quanteda::tokens()} or to the
-user-specified modeling function}
-
 \item{tolower}{logical. Convert all text to lowercase? If \code{model} is an
 embeddings object, this value is passed to \code{\link[quanteda:dfm]{quanteda::dfm()}}.}
+
+\item{output_embeddings}{\code{FALSE} (the default) returns a tibble.
+\code{TRUE} returns an embeddings object. See 'Value' for details.}
 
 \item{text_col}{string. a column of texts for which to compute embeddings}
 
 \item{id_col}{optional string. column of unique document ids}
 
 \item{.keep_all}{logical. Keep all columns from input? Ignored if \code{output_embeddings = TRUE}.}
+
+\item{dfm}{a quanteda \link[quanteda:dfm]{dfm}}
 }
 \description{
 \code{textstat_embedding()} takes a 'quanteda' \link[quanteda:dfm]{dfm}. \code{embed_docs()} is a more
