### 2025-08-06 9da0cf7
tests + documentation reflect recent changes
