### 2025-08-06 2b57568
consistency in linking to reference documentation
addresses issue #9
