+++ README.Rmd
@@ -188,7 +188,7 @@ new_with_valence_rotation
 
 #### Align and Compare Embeddings Models
 
-Aligning separately trained embeddings can be useful for tracking semantic change over time or semantic differences between training datasets. It can also be useful for comparing texts across languages. `align_embeddings(x, y)` rotates the embeddings in `x` (and changes their dimensionality if necessary) so that they can be compared with those in `y`. Optionally, `matching` can be used to specify one-to-one matching between  embeddings in the two models (e.g. a bilingual dictionary).
+Aligning separately trained embeddings can be useful for tracking semantic change over time or semantic differences between training datasets. It can also be useful for comparing texts across languages. `align_embeddings(x, y)` rotates the embeddings in `x` (and changes their dimensionality if necessary) so that they can be compared with those in `y`. Optionally, `matching` can be used to specify one-to-one matching between  embeddings in the two models (e.g. a bilingual dictionary). See `align_embeddings()` for full documentation.
 
 Once aligned, groups of embeddings can be compared using `total_dist()` or `average_sim()`. These can be used to quantify the overall distance or similarity between two parallel embedding spaces.
 
