### 2025-08-06 385af26
clarify package name
addresses issue #10
