+++ paper/paper.bib
@@ -329,6 +329,14 @@ The book is published at ds4psych.com},
 	file = {Full Text PDF:/Users/louisteitelbaum/Zotero/storage/J935AK82/Wickham et al. - 2019 - Welcome to the Tidyverse.pdf:application/pdf},
 }
 
+@Manual{wickham2025,
+  title = {dplyr: A Grammar of Data Manipulation},
+  author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller and Davis Vaughan},
+  year = {2025},
+  note = {R package version 1.1.4},
+  url = {https://dplyr.tidyverse.org},
+}
+
 @article{feuerriegel2025,
 	title = {Using natural language processing to analyse text data in behavioural science},
 	copyright = {2025 Springer Nature America, Inc.},
