+++ paper/paper.md
@@ -26,7 +26,7 @@ journal: JOSS
 
 # Statement of Need
 
-Dense vector embeddings are the fundamental building block of modern natural language processing [@lauriola2022]. The ability to represent the meaning of texts as a continuous, high dimensional space underlies the recent successes of large language models (LLMs). Embeddings have also revolutionized research methods and inspired new theoretical frameworks in linguistics, psychology, sociology, and neuroscience [see e.g. @duran2019; @feuerriegel2025; @grand2022; @hamilton2018; @kjell2023; @kozlowski2019; @schrimpf2021]. As text embeddings become ubiquitous in the social and behavioral sciences, the need for flexible, easy-to-learn tools increases. Answering this need, **embedplyr** (pronounced "embe-DEE-plier") enables common operations with word and text embeddings within a **tidyverse** [@wickham2019] and/or **quanteda** [@benoit2018] workflow, as demonstrated in @teitelbaum2024. **embedplyr** is designed to facilitate the use of word and text embeddings in common data manipulation and text analysis workflows, without introducing new syntax or unfamiliar data structures to R users.
+Dense vector embeddings are the fundamental building block of modern natural language processing [@lauriola2022]. The ability to represent the meaning of texts as a continuous, high dimensional space underlies the recent successes of large language models (LLMs). Embeddings have also revolutionized research methods and inspired new theoretical frameworks in linguistics, psychology, sociology, and neuroscience [see e.g. @duran2019; @feuerriegel2025; @grand2022; @hamilton2018; @kjell2023; @kozlowski2019; @schrimpf2021]. As text embeddings become ubiquitous in the social and behavioral sciences, the need for flexible, easy-to-learn tools increases. Answering this need, **embedplyr** (pronounced "embe-DEE-plier") enables common operations with word and text embeddings within a **tidyverse** [@wickham2019] and/or **quanteda** [@benoit2018] workflow, as demonstrated in @teitelbaum2024. Much as **dplyr** is "a grammar of data manipulation, providing a consistent set of verbs that help you solve the most common data manipulation challenges" [@wickham2025], **embedplyr** is a grammar of embeddings manipulation, designed to facilitate the use of word and text embeddings in common analysis workflows without introducing new syntax or unfamiliar data structures to R users.
 
 Existing tools for working with embeddings in R are generally specific to particular model architectures. For example, **word2vec** [@wijffels2023] provides access to word2vec models [@mikolov2013a; @mikolov2013b], **text2vec** [@selivanov2023] provides access to LDA, LSA, and GloVe models [@blei2003; @deerwester1990; @pennington2014], **fastText** [@mouselimis2024] provides access to fastText models [@facebook2016; @bojanowski2017a; @bojanowski2017b; @grave2018], and **text** [@kjell2023] provides access to transformers-based LLMs [@wolf2020]. While these tools are already invaluable in the social and behavioral sciences, using different syntax and separate data structures for each model architecture can be cumbersome. This is especially true given the prevalence of studies that make use of multiple model architectures in parallel analyses [e.g. @carrella2023; @hussain2024; @markus2024]. **embedplyr** fills this gap with a model agnostic approach; it can be used to work with embeddings from any model framework using syntax that will be familiar to any **tidyverse** user. While some existing tools focus on convenience functions for encouraging expert-recommended best practices [e.g. @kjell2023], **embedplyr** prioritizes flexibility---much like its namesake, **dplyr** [@wickham2023]. This approach yields simple, modular functions that are useful both for educating students [see @teitelbaum2024] and experimenting with novel methods.
 
