### 2025-08-06 rimonim/embedplyr#10 @rimonim
Yes you're right I should have explained this more clearly.
In dplyr, [the d is for dataframes, the plyr is to evoke pliers](https://github.com/tidyverse/dplyr/issues/1857). embedplyr is the same but with embeddings instead of dataframes. To be clear, the intention is not to suggest that embedplyr is a tidyverse package. On the contrary, I don't think that tidy tibbles are particularly well suited to work with embeddings (hence the embeddings object class). What I'd really like to convey with the name is the idea of a flexible grammar, which makes common operations easy but also enables less common ones with a similar degree of transparency and readability.
I added a sentence to the paper to clarify all this.
