+++ pkgdown.yml
@@ -1,9 +1,10 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   DDR: DDR.html
-last_built: 2025-03-24T08:52Z
+  performance: performance.html
+last_built: 2025-08-06T19:24Z
 urls:
   reference: https://rimonim.github.io/embedplyr/reference
   article: https://rimonim.github.io/embedplyr/articles
