### 2025-08-22 rimonim/embedplyr#6 @cmaimone
I get the argument for having a function for this - I just personally would pick something other than overloading predict.  This one's really up to you though - just wanted to raise it for consideration.
