### 2025-09-01 e294ccb
change `predict.embeddings()` to `emb()`
addresses issue #6
