+++ paper/paper.md
@@ -53,11 +53,11 @@ eng.fiction.all_sgns.1800 <- load_embeddings(
 	)
 ```
 
-The output of `load_embeddings()` is an embeddings object. An embeddings object is simply a numeric matrix with fast hash table indexing by rownames (generally tokens). This means that it can be easily coerced to a dataframe or tibble, while also allowing special embeddings-specific methods and functions, such as `predict.embeddings()` and `find_nearest()`:
+The output of `load_embeddings()` is an embeddings object. An embeddings object is simply a numeric matrix with fast hash table indexing by rownames (generally tokens). This means that it can be easily coerced to a dataframe or tibble, while also allowing special embeddings-specific methods and functions, such as `emb()` and `find_nearest()`:
 
 
 ```r
-moral_embeddings <- predict(glove_twitter_25d, c("good", "bad"))
+moral_embeddings <- emb(glove_twitter_25d, c("good", "bad"))
 moral_embeddings
 ```
 
@@ -127,8 +127,8 @@ high_anx_dict <- c("anxious", "overwhelmed", "nervous", "stressed")
 low_anx_dict <- c("relaxed", "calm", "mellow")
 
 # embed dictionaries
-high_anx_dict_embeddings <- predict(glove_twitter_25d, high_anx_dict)
-low_anx_dict_embeddings <- predict(glove_twitter_25d, low_anx_dict)
+high_anx_dict_embeddings <- emb(glove_twitter_25d, high_anx_dict)
+low_anx_dict_embeddings <- emb(glove_twitter_25d, low_anx_dict)
 
 # average embeddings to create DDR
 high_anx_DDR <- average_embedding(high_anx_dict_embeddings)
