+++ R/embed_tokens.R
@@ -62,7 +62,7 @@ embed_tokens.default <- function(x, model, ..., .keep_missing = FALSE,
 	if (any(embedding_not_found)) {
 		warning(sprintf("%d tokens in `x` are not present in `model`.", sum(embedding_not_found)))
 	}
-	out <- lapply(x, function(nd) suppressWarnings(predict.embeddings(model, nd, drop = FALSE, .keep_missing = .keep_missing)))
+	out <- lapply(x, function(nd) suppressWarnings(emb(model, nd, drop = FALSE, .keep_missing = .keep_missing)))
 	if (output_embeddings) return(out)
 	out <- lapply(out, function(emb) if (nrow(emb) != 0) as_tibble(emb, rownames = "token") else NULL)
 	dplyr::bind_rows(out, .id = "doc_id")
