+++ R/sim_matrix.R
@@ -34,7 +34,7 @@
 #'
 #' @import tibble
 #' @examples
-#' emb <- predict(glove_twitter_25d, c("table", "chair", "cat"))
+#' emb <- emb(glove_twitter_25d, c("table", "chair", "cat"))
 #'
 #' sim_matrix(emb)
 #' sim_matrix(emb, method = "euclidean")
