+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   DDR: DDR.html
   performance: performance.html
-last_built: 2025-08-06T19:24Z
+last_built: 2025-09-01T10:36Z
 urls:
   reference: https://rimonim.github.io/embedplyr/reference
   article: https://rimonim.github.io/embedplyr/articles
