### 2025-09-29 0358b5e
Refrence contributor guidelines in README
addresses JOSS reviewer feedback #1
