+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   DDR: DDR.html
   performance: performance.html
-last_built: 2025-09-01T10:36Z
+last_built: 2025-09-29T05:23Z
 urls:
   reference: https://rimonim.github.io/embedplyr/reference
   article: https://rimonim.github.io/embedplyr/articles
