### 2025-09-29 a984452
lifecycle badges
predict.embeddings() deprecated in favor of emb()
