+++ R/emb.R
@@ -1,6 +1,6 @@
 #' Retrieve Token Embeddings
 #'
-#' @param object an embeddings object made by `load_embeddings()` or `as.embeddings()`
+#' @param x an embeddings object made by `load_embeddings()` or `as.embeddings()`
 #' @param newdata a character vector of tokens
 #' @param drop logical. If `TRUE` (the default) and the result is one-dimensional
 #' (e.g. a single row), the output will be a (named) vector.
@@ -11,7 +11,7 @@
 #'
 #' @details
 #' Duplicated items in `newdata` will result in duplicated rows in the output.
-#' If an item in `newdata` matches multiple rows in `object`, the last one will
+#' If an item in `newdata` matches multiple rows in `x`, the last one will
 #' be returned.
 #'
 #' @section Value:
@@ -25,23 +25,37 @@
 #' texts_embeddings
 
 #' @export
-emb <- function(object, newdata, drop = TRUE, .keep_missing = FALSE){
+emb <- function(x, newdata, drop = TRUE, .keep_missing = FALSE){
+  if (!is.embeddings(x)) stop("x must be an embeddings object.")
   if (any(zchars <- !nzchar(newdata))) {
     warning(sprintf('Replacing %d empty strings with " ".', sum(zchars)))
     newdata[zchars] <- " "
   }
-  embedding_not_found <- !vapply(newdata, exists, FALSE, envir = attr(object, "token_index"))
+  embedding_not_found <- !vapply(newdata, exists, FALSE, envir = attr(x, "token_index"))
   if (any(embedding_not_found)) {
     warning(sprintf("%d items in `newdata` are not present in the embeddings object.", sum(embedding_not_found)))
   }
   if (.keep_missing) {
-    out <- matrix(nrow = length(newdata), ncol = ncol(object), dimnames = list(newdata, colnames(object)))
-    out[!embedding_not_found,] <- object[newdata[!embedding_not_found],]
+    out <- matrix(nrow = length(newdata), ncol = ncol(x), dimnames = list(newdata, colnames(x)))
+    out[!embedding_not_found,] <- x[newdata[!embedding_not_found],]
     out <- as.embeddings(out, rowname_repair = FALSE)
     if (drop && nrow(out) == 1L) return(out[1,])
   }else{
-    out <- object[newdata[!embedding_not_found],,drop = drop]
+    out <- x[newdata[!embedding_not_found],,drop = drop]
     if (drop && length(out) == 0L) return(numeric())
   }
   out
+}
+
+#' Retrieve Token Embeddings
+#'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' `predict.embeddings()` was renamed to `emb()` to create a more intuitive API.
+#' @keywords internal
+#' @export
+predict.embeddings <- function(object, newdata, drop = TRUE, .keep_missing = FALSE){
+  lifecycle::deprecate_warn("1.0.0", "predict.embeddings()", "emb()")
+  emb(x = object, newdata = newdata, drop = drop, .keep_missing = .keep_missing)
 }
