+++ man/predict.embeddings.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/emb.R
+\name{predict.embeddings}
+\alias{predict.embeddings}
+\title{Retrieve Token Embeddings}
+\usage{
+\method{predict}{embeddings}(object, newdata, drop = TRUE, .keep_missing = FALSE)
+}
+\description{
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
+
+\code{predict.embeddings()} was renamed to \code{emb()} to create a more intuitive API.
+}
+\keyword{internal}
