+++ NAMESPACE
@@ -27,6 +27,7 @@ S3method(normalize,default)
 S3method(normalize,numeric)
 S3method(normalize_rows,data.frame)
 S3method(normalize_rows,default)
+S3method(predict,embeddings)
 S3method(print,embeddings)
 S3method(rbind,embeddings)
 S3method(reduce_dimensionality,data.frame)
@@ -65,6 +66,7 @@ export(textstat_embedding)
 export(total_dist)
 import(quanteda)
 import(tibble)
+importFrom(lifecycle,deprecated)
 importFrom(pillar,style_subtle)
 importFrom(rlang,"%||%")
 importFrom(rlang,":=")
