### 2025-10-20 71e06b0
references formatting
