+++ paper/paper.bib
@@ -181,7 +181,7 @@ Publisher: American Psychological Association},
 }
 
 @article{benoit2018,
-	title = {quanteda: {An} {R} package for the quantitative analysis of textual data},
+	title = {{quanteda}: {An} {R} package for the quantitative analysis of textual data},
 	volume = {3},
 	issn = {2475-9066},
 	shorttitle = {quanteda},
@@ -407,7 +407,6 @@ The book is published at ds4psych.com},
 	issn = {1532-4435},
 	doi = {10.7551/mitpress/1120.003.0082},
 	abstract = {We describe latent Dirichlet allocation (LDA), a generative probabilistic model for collections of discrete data such as text corpora. LDA is a three-level hierarchical Bayesian model, in which each item of a collection is modeled as a finite mixture over an underlying set of topics. Each topic is, in turn, modeled as an infinite mixture over an underlying set of topic probabilities. In the context of text modeling, the topic probabilities provide an explicit representation of a document. We present efficient approximate inference techniques based on variational methods and an EM algorithm for empirical Bayes parameter estimation. We report results in document modeling, text classification, and collaborative filtering, comparing to a mixture of unigrams model and the probabilistic LSI model.},
-	number = {null},
 	journal = {J. Mach. Learn. Res.},
 	author = {Blei, David M. and Ng, Andrew Y. and Jordan, Michael I.},
 	month = mar,
@@ -418,8 +417,8 @@ The book is published at ds4psych.com},
 
 @inproceedings{pennington2014,
 	address = {Doha, Qatar},
-	title = {Glove: {Global} {Vectors} for {Word} {Representation}},
-	shorttitle = {Glove},
+	title = {{GloVe}: {Global} {Vectors} for {Word} {Representation}},
+	shorttitle = {GloVe},
 	url = {http://aclweb.org/anthology/D14-1162},
 	doi = {10.3115/v1/D14-1162},
 	abstract = {Recent methods for learning vector space representations of words have succeeded in capturing ﬁne-grained semantic and syntactic regularities using vector arithmetic, but the origin of these regularities has remained opaque. We analyze and make explicit the model properties needed for such regularities to emerge in word vectors. The result is a new global logbilinear regression model that combines the advantages of the two major model families in the literature: global matrix factorization and local context window methods. Our model efﬁciently leverages statistical information by training only on the nonzero elements in a word-word cooccurrence matrix, rather than on the entire sparse matrix or on individual context windows in a large corpus. The model produces a vector space with meaningful substructure, as evidenced by its performance of 75\% on a recent word analogy task. It also outperforms related models on similarity tasks and named entity recognition.},
@@ -430,7 +429,7 @@ The book is published at ds4psych.com},
 	author = {Pennington, Jeffrey and Socher, Richard and Manning, Christopher},
 	year = {2014},
 	pages = {1532--1543},
-	file = {Pennington et al. - 2014 - Glove Global Vectors for Word Representation.pdf:/Users/louisteitelbaum/Zotero/storage/7CX7JWZX/Pennington et al. - 2014 - Glove Global Vectors for Word Representation.pdf:application/pdf},
+	file = {Pennington et al. - 2014 - GloVe Global Vectors for Word Representation.pdf:/Users/louisteitelbaum/Zotero/storage/7CX7JWZX/Pennington et al. - 2014 - Glove Global Vectors for Word Representation.pdf:application/pdf},
 }
 
 @inproceedings{wolf2020,
[…]
