### 2025-10-23 2db78cd
Increment version number to 0.3.0
