### 2025-10-23 #8262 @rimonim
- [x] Author and affiliation are correct
- [x] New release version number: [0.3.0](https://github.com/rimonim/embedplyr/releases/tag/v0.3.0)
- [x] Zenodo archive DOI: [10.5281/zenodo.17426404](https://doi.org/10.5281/zenodo.17426404)
- [x] title, author, and license match JOSS paper + Github repo
