+++ llms.txt
@@ -1,0 +1,456 @@
+# embedplyr
+
+## Overview
+
+embedplyr enables common operations with word and text embeddings within
+a ‘tidyverse’ and/or ‘quanteda’ workflow, as demonstrated in [Data
+Science for Psychology: Natural Language](http://ds4psych.com).
+
+- [`load_embeddings()`](https://rimonim.github.io/embedplyr/reference/load_embeddings.md)
+  loads pretrained [GloVe](https://nlp.stanford.edu/projects/glove/),
+  [word2vec](https://code.google.com/archive/p/word2vec/),
+  [HistWords](https://nlp.stanford.edu/projects/histwords/), [ConceptNet
+  Numberbatch](https://github.com/commonsense/conceptnet-numberbatch),
+  and [fastText](https://fasttext.cc) word embedding models from
+  Internet sources or from your working directory
+- [`embed_tokens()`](https://rimonim.github.io/embedplyr/reference/embed_tokens.md)
+  returns the embedding for each token in a set of texts
+- [`embed_docs()`](https://rimonim.github.io/embedplyr/reference/embed_docs.md)
+  generates text embeddings for a set of documents
+- [`get_sims()`](https://rimonim.github.io/embedplyr/reference/get_sims.md)
+  calculates row-wise similarity metrics between a set of embeddings and
+  a given reference
+- [`average_embedding()`](https://rimonim.github.io/embedplyr/reference/average_embedding.md)
+  calculates the (weighted) average of multiple embeddings
+- [`reduce_dimensionality()`](https://rimonim.github.io/embedplyr/reference/reduce_dimensionality.md)
+  reduces the dimensionality of embeddings
+- [`align_embeddings()`](https://rimonim.github.io/embedplyr/reference/align_embeddings.md)
+  rotates the embeddings from one model so that they can be compared
+  with those from another
+- [`normalize()`](https://rimonim.github.io/embedplyr/reference/normalize.md)
+  and
+  [`normalize_rows()`](https://rimonim.github.io/embedplyr/reference/normalize.md)
+  normalize embeddings to the unit hypersphere
+- and more…
+
+## Installation
+
+You can install the development version of embedplyr from
+[GitHub](https://github.com/) with:
+
+``` r
+remotes::install_github("rimonim/embedplyr") 
+```
+
+## Functionality
+
+embedplyr is designed to facilitate the use of word and text embeddings
+in common data manipulation and text analysis workflows, without
+introducing new syntax or unfamiliar data structures.
+
+embedplyr is model agnostic; it can be used to work with embeddings from
+decontextualized models like
+[GloVe](https://nlp.stanford.edu/projects/glove/) and
+[word2vec](https://code.google.com/archive/p/word2vec/), or from
+contextualized models like BERT or others made available through the
+‘[text](https://r-text.org)’ package.
+
+``` r
+library(embedplyr)
+```
+
+### Loading Pretrained Embeddings
+
+embedplyr won’t help you train new embedding models, but it can load
+embeddings from a file or download them from online. This is especially
+useful for pretrained word embedding models like GloVe, word2vec, and
[…]
