+++ reference/token_index_add.md
@@ -1,0 +1,19 @@
+# Update hash table after additions
+
+Update hash table after additions
+
+## Usage
+
+``` r
+token_index_add(embeddings, added_tokens)
+```
+
+## Arguments
+
+- embeddings:
+
+  an embeddings object
+
+- added_tokens:
+
+  tokens to be added to the hash table

[…]
